Fix CI: armv7/aarch64 cross-compile, clippy unknown lint
- Fix aarch64 cross-compile: use libc::c_char instead of i8 for
  sockaddr_un.sun_path (differs between u8/i8 on aarch64 vs x86_64)
- Gate seccomp module behind target_arch = x86_64 (BPF is arch-specific)
- Fix armv7 cross-compile: add type casts for 32-bit time_t/i32 types
- Remove unknown clippy::unused_assignments lint
- Regenerate port-parity.md with dynamic test counting
- All CI checks pass: clippy (0 errors), test (1345 pass),
  cross-compile (aarch64, armv7), freshness gate, no-orig, self-test oracle (299/299)

+++ crates/openntpd-rs-io/src/clock.rs
@@ -159,7 +159,7 @@ pub fn adjfreq(freq: openntpd_rs_core::util::Frequency) -> ClockResult<()> {
     let linux_freq = openbsd_freq_to_linux(freq)?;
     let mut tx: libc::timex = unsafe { std::mem::zeroed() };
     tx.modes = libc::ADJ_FREQUENCY;
-    tx.freq = linux_freq;
+    tx.freq = linux_freq as libc::c_long;
     // SAFETY: adjtimex is safe; timex is fully initialized.
     let ret = unsafe { libc::adjtimex(&mut tx) };
     if ret == -1 {
